requirements.txt
ninja  # For faster builds.
psutil
ray >= 2.5.1
pandas  # Required for Ray data.
pyarrow  # Required for Ray data.
sentencepiece  # Required for LLaMA tokenizer.
numpy
torch >= 2.0.0
transformers >= 4.33.1  # Required for Code Llama.
xformers >= 0.0.22
fastapi
uvicorn[standard]
pydantic < 2  # Required for OpenAI server.
gurobipy
rich
deepspeed == 0.12.3
deepspeed-kernels